refactor: remove dead config_store module and split config into lifecycle-based submodules
